CEDR: Factor tree namespace out of qlt namespace.
The tree can be used in other applications, e.g., a tree-based BFB reduction
with mixed reduction operations.
